centralize colors class
