+++ analysis/prompt_rl_optimizer.py
@@ -4,41 +4,13 @@ from __future__ import annotations
 
 import random
 import logging
-import re # For cleaning generated text (though primarily used in base PromptOptimizer, good to have)
+import re  # For cleaning generated text (though primarily used in base PromptOptimizer, good to have)
 from typing import Callable, Dict, Optional, List, Any
 
 # These imports are assumed to work correctly from the user's setup as per the last instruction
-from .prompt_optimizer import PromptOptimizer # Base class providing LLM, scoring, safety, and variation generation
+from .prompt_optimizer import PromptOptimizer  # Base class providing LLM, scoring, safety, and variation generation
 
-
-# --- ANSI Escape Codes for Colors and Styles (for clear console output) ---
-class Colors:
-    RESET = "\033[0m"
-    BOLD = "\033[1m"
-    DIM = "\033[2m"
-    ITALIC = "\033[3m"
-    UNDERLINE = "\033[4m"
-    BLINK = "\033[5m"
-    INVERT = "\033[7m"
-    HIDDEN = "\033[8m"
-    STRIKETHROUGH = "\033[9m"
-
-    BLACK = "\033[30m"
-    RED = "\033[31m"
-    GREEN = "\033[32m"
-    YELLOW = "\033[33m"
-    BLUE = "\033[34m"
-    MAGENTA = "\033[35m"
-    CYAN = "\033[36m"
-    WHITE = "\033[37m"
-
-    BRIGHT_BLACK = "\033[90m"
-    BRIGHT_RED = "\033[91m"
-    BRIGHT_GREEN = "\033[92m"
-    BRIGHT_YELLOW = "\033[93m"
-    BRIGHT_BLUE = "\033[94m"
-    BRIGHT_MAGENTA = "\033[95m"
-    BRIGHT_CYAN = "\033[96m"
+from utils.colors import Colors
 
 # --- Configure Logging for structured and colored output ---
 class ColoredFormatter(logging.Formatter):
