+++ research_workflow/topic_selector.py
@@ -4,39 +4,12 @@ from __future__ import annotations
 
 import logging
 from typing import Optional, Any, List
-import re # For parsing generated text
+import re  # For parsing generated text
 
 import torch
 from transformers import AutoModelForCausalLM, AutoTokenizer, GenerationConfig, pipeline
 
-# --- ANSI Escape Codes for Colors and Styles ---
-class Colors:
-    RESET = "\033[0m"
-    BOLD = "\033[1m"
-    DIM = "\033[2m"
-    ITALIC = "\033[3m"
-    UNDERLINE = "\033[4m"
-    BLINK = "\033[5m"
-    INVERT = "\033[7m"
-    HIDDEN = "\033[8m"
-    STRIKETHROUGH = "\033[9m"
-
-    BLACK = "\033[30m"
-    RED = "\033[31m"
-    GREEN = "\033[32m"
-    YELLOW = "\033[33m"
-    BLUE = "\033[34m"
-    MAGENTA = "\033[35m"
-    CYAN = "\033[36m"
-    WHITE = "\033[37m"
-
-    BRIGHT_BLACK = "\033[90m"
-    BRIGHT_RED = "\033[91m"
-    BRIGHT_GREEN = "\033[92m"
-    BRIGHT_YELLOW = "\033[93m"
-    BRIGHT_BLUE = "\033[94m"
-    BRIGHT_MAGENTA = "\033[95m"
-    BRIGHT_CYAN = "\033[96m"
+from utils.colors import Colors
 
 # Configure a basic logger for the module
 log = logging.getLogger(__name__)
